Multi server

+++ experiment.py
@@ -1,31 +1,29 @@
 """QUIC Experiment Harness
 
 Usage:
-    experiment.py experiment.py [--device DEVICE] [--options OPTIONS] [--browsers BROWSERS] [--url URL] [--runs RUNS] [options]
+    experiment.py experiment.py [--device DEVICE] [--options OPTIONS ...] [--browsers BROWSERS ...] [--url URL] [--runs RUNS] [options] 
     
 Arguments:
     --device DEVICE           Network device to modify [default: lo root]
     --options OPTIONS         tc-netem conditions to apply [default: delay 0ms]
-    --browsers BROWSERS       List of browsers to test [default: firefox,chromium,edge]
+    --browsers BROWSERS       List of browsers to test [default: firefox chromium edge]
     --url URL                 URL to access [default: https://localhost]
     --runs RUNS               Number of runs in the experiment [default: 1]
 
 Options:
     -h --help                 Show this screen 
     --disable_caching         Disables caching
+    --multi-server            Uses all four server ports
     --warmup                  Warms up connection
 """
+
+import os, cache_control, time, random, subprocess, csv, json
 from typing import List
-import subprocess, csv, json
-import os
 from playwright import sync_playwright
 from docopt import docopt
-import cache_control
-import time
-import random
 from getTime import getTime
-
 from args import getArguments
+from tqdm import tqdm
 
 
 # generated command line code
@@ -37,6 +35,7 @@ experimentParameters = [
     "experimentStartTime",
     "netemParams", # TODO: think about better encoding
     "httpVersion", 
+    "server",
     "warmup",
 ]
 
@@ -67,7 +66,10 @@ def main():
     args = docopt(__doc__, argv=None, help=True, version=None, options_first=False)
     device = args['--device']
     options = args['--options']
-    browsers = args['--browsers'].split(",")
+    # Fix docopt splitting default argument
+    if options == ["delay", "0ms"]:
+        options = ["delay 0ms"]
+    browsers = args['--browsers']
     url = args['--url']
     runs = int(args['--runs'])
 
@@ -75,46 +77,60 @@ def main():
     warmup_connection = args['--warmup']
 
     server_conf = "/usr/local/nginx/conf/nginx.conf"
-    
+
     if disable_caching:
         # Assumes that there is server caching by default
         cache_control.remove_server_caching(server_conf, 23)
     # Make sure server is running
-    subprocess.run("sudo systemctl restart nginx.service".split())
+    if "localhost" in url or "127.0.0.1" in url:
+        subprocess.run("sudo systemctl restart nginx.service".split())
+    with sync_playwright() as p:
+        for netemParams in tqdm(options, desc="Experiments"):
+            reset = RESET_FORMAT.format(DEVICE=device)
 
-    reset = RESET_FORMAT.format(DEVICE=device)
+            call  = CALL_FORMAT.format(DEVICE=device, OPTIONS=netemParams)
+            experimentID = int(time.time()) # ensures no repeats
+            for browser in tqdm(browsers, f"Browsers for '{netemParams}'"):
+                name = browser + "_" + netemParams.replace(" ", "_")
+                directoryPath = "results"
+                csvFileName = f"{directoryPath}/{name}.csv"
 
-    call  = CALL_FORMAT.format(DEVICE=device, OPTIONS=options)
-    experimentID = int(time.time()) # ensures no repeats
-    netemParams = options
-    
-    for browser in browsers:
-        name = browser + "_" + options.replace(" ", "_")
-        directoryPath = "results"
-        csvFileName = f"{directoryPath}/{name}.csv"
-
-        # Setup data file headers
-        os.makedirs(os.path.dirname(csvFileName), exist_ok=True)
-        with open(csvFileName, 'w', newline='\n') as outFile:
-            csvWriter = csv.writer(outFile)
-            csvWriter.writerow(parameters)
-        
-        whenRunH3 = runs * [True] + runs * [False]
-        random.shuffle(whenRunH3)
-
-        # run the same experiment multiple times over h3/h2
-        with sync_playwright() as p:
-            for useH3 in whenRunH3:
-                results = runExperiment(call, reset, p, browser, useH3, url, warmup=warmup_connection)
-                results["experimentID"] = experimentID
-                results["netemParams"] = netemParams
-                results["httpVersion"] = "h3" if useH3 else "h2"
-                results["warmup"] = warmup_connection
-                writeData(results, csvFileName)
-
-    if disable_caching:
+                # Setup data file headers
+                os.makedirs(os.path.dirname(csvFileName), exist_ok=True)
+                if not os.path.exists(csvFileName):
+                    with open(csvFileName, 'w', newline='\n') as outFile:
+                        csvWriter = csv.writer(outFile)
+                        csvWriter.writerow(parameters)
+                
+                whenRunH3 = runs * [True] + runs * [False]
+                random.shuffle(whenRunH3)
+                perServer = runs // 4
+                # Ensure all servers are represented the same amount in H2 vs. H3
+                if runs < 4 or not args['--multi-server']:
+                    whichServer = [':443'] * (runs * 2)
+                else:
+                    servers1 = [':443', ':444', ':445', ':7080/login.php'] * perServer
+                    servers2 = [':443', ':444', ':445', ':7080/login.php'] * perServer
+                    random.shuffle(servers1)
+                    random.shuffle(servers2)
+                    whichServer = servers1 + servers2
+                # run the same experiment multiple times over h3/h2
+                for useH3 in tqdm(whenRunH3, desc=f"Runs for {browser}"):
+                    results = runExperiment(call, reset, p, browser, useH3, url, whichServer.pop(), warmup=warmup_connection)
+                    results["experimentID"] = experimentID
+                    results["netemParams"] = netemParams
+                    results["httpVersion"] = "h3" if useH3 else "h2"
+                    results["warmup"] = warmup_connection
+                    writeData(results, csvFileName)
+                    httpVersion = "HTTP/3" if useH3 else "HTTP/2"
+                    # Print info from latest run and then go back lines to prevent broken progress bars
+                    tqdm.write(f"\033[F\033[K{browser}: {results['server']} ({httpVersion})       ")
+                print("", end="\033[F\033[K")
+            print("", end="\033[F\033[K")
+    if args['--disable_caching']:
         # Re-enable server caching
         cache_control.add_server_caching(server_conf, 23, 9)
+    print("Finished!\n")
 
 def writeData(data: json, csvFileName: str):
     with open(csvFileName, 'a+', newline='\n') as outFile:
@@ -136,19 +152,21 @@ def launchBrowser(
     browserType: str,
     url: str, 
     h3: bool,
+    port: str,
     warmup: bool,
 ) -> json:
     if browserType  ==  "firefox":
-        return launchFirefox(pwInstance, url, h3, warmup)
+        return launchFirefox(pwInstance, url, h3, port, warmup)
     elif browserType  ==  "chromium":
-        return launchChromium(pwInstance, url, h3, warmup)
+        return launchChromium(pwInstance, url, h3, port, warmup)
     elif browserType  ==  "edge":
-        return launchEdge(pwInstance, url, h3, warmup)
+        return launchEdge(pwInstance, url, h3, port, warmup)
 
 def launchFirefox(
     pwInstance: "SyncPlaywrightContextManager", 
     url: str, 
     h3: bool,
+    port: str,
     warmup: bool,
 ) -> json:
     firefoxPrefs = {}
@@ -157,30 +175,19 @@ def launchFirefox(
     if h3:
         domain = url if "https://" not in url else url[8:]
         firefoxPrefs["network.http.http3.enabled"] = True
-        firefoxPrefs["network.http.http3.alt-svc-mapping-for-testing"] = f"{domain};h3-29=:443"
+        firefoxPrefs["network.http.http3.alt-svc-mapping-for-testing"] = f"{domain};h3-29={port.split('/')[0]}"
 
     browser = pwInstance.firefox.launch(
         headless=True,
         firefoxUserPrefs=firefoxPrefs,
     )
-    context = browser.newContext()
-    page = context.newPage()
-    warmupIfSpecified(page, url, warmup)
-    response = page.goto(url)
-    print("Firefox: ", response.headers['version'])
-
-    # getting performance timing data
-    # if we don't stringify and parse, things break
-    timingFunction = '''JSON.stringify(window.performance.getEntriesByType("navigation")[0])'''
-    performanceTiming = json.loads(page.evaluate(timingFunction))
-
-    browser.close()
-    return performanceTiming
+    return getResults(browser, url, h3, port, warmup)
 
 def launchChromium(
     pwInstance: "SyncPlaywrightContextManager", 
     url: str, 
     h3: bool,
+    port: str,
     warmup: bool,
 ) -> json:
     chromiumArgs = []
@@ -188,54 +195,65 @@ def launchChromium(
         domain = url if "https://" not in url else url[8:]
         chromiumArgs = ["--enable-quic", f"--origin-to-force-quic-on={domain}:443", "--quic-version=h3-29"]
 
-    browser =  pwInstance.chromium.launch(
-        headless=True,
-        args=chromiumArgs,
-    )
-    context = browser.newContext()
-    context = browser.newContext()
-    page = context.newPage()
-    warmupIfSpecified(page, url, warmup)
-    response = page.goto(url)
-    print("Chromium: ",response.headers['version'])
-
-    # getting performance timing data
-    # if we don't stringify and parse, things break
-    timingFunction = '''JSON.stringify(window.performance.getEntriesByType("navigation")[0])'''
-    performanceTiming = json.loads(page.evaluate(timingFunction))
-    
-    browser.close()
-    return performanceTiming
+    try:
+        browser =  pwInstance.chromium.launch(
+            headless=True,
+            args=chromiumArgs,
+        )
+    except:
+        browser =  pwInstance.chromium.launch(
+            headless=True,
+            args=chromiumArgs,
+        )
+    return getResults(browser, url, h3, port, warmup)
 
 def launchEdge(
     pwInstance: "SyncPlaywrightContextManager", 
     url: str, 
     h3: bool,
+    port: str,
     warmup: bool,
 ) -> json:
     edgeArgs = []
     if (h3) :
         domain = url if "https://" not in url else url[8:]
         edgeArgs = ["--enable-quic", f"--origin-to-force-quic-on={domain}:443", "--quic-version=h3-29"]
-    browser = pwInstance.chromium.launch(
-        headless=True,
-        executablePath='/opt/microsoft/msedge-dev/msedge',
-        args=edgeArgs,
-    )
+    try:
+        browser = pwInstance.chromium.launch(
+            headless=True,
+            executablePath='/opt/microsoft/msedge-dev/msedge',
+            args=edgeArgs,
+        )
+    except:
+        browser = pwInstance.chromium.launch(
+            headless=True,
+            executablePath='/opt/microsoft/msedge-dev/msedge',
+            args=edgeArgs,
+        )
+    return getResults(browser, url, h3, port, warmup)
+    
+
+def getResults (
+    browser,
+    url: str, 
+    h3: bool,
+    port: str,
+    warmup: bool,
+) -> json:
     context = browser.newContext()
     page = context.newPage()
-    warmupIfSpecified(page, url, warmup)
-    response = page.goto(url)
-    print("Edge: ",response.headers['version'])
+    warmupIfSpecified(page, url + port, warmup)
+    response = page.goto(url + port)
 
     # getting performance timing data
     # if we don't stringify and parse, things break
     timingFunction = '''JSON.stringify(window.performance.getEntriesByType("navigation")[0])'''
     performanceTiming = json.loads(page.evaluate(timingFunction))
+    performanceTiming['server'] = response.headers['server']
     
     browser.close()
     return performanceTiming
-    
+
 def runExperiment(
     call: str, 
     reset: str, 
@@ -243,10 +261,11 @@ def runExperiment(
     browserType: str, 
     h3: bool,
     url: str,
+    port: str,
     warmup: bool,
 ) -> json:
     runTcCommand(call)
-    results = launchBrowser(pwInstance, browserType, url, h3, warmup)
+    results = launchBrowser(pwInstance, browserType, url, h3, port, warmup=warmup)
     runTcCommand(reset)
 
     return results
@@ -255,12 +274,13 @@ def runExperiment(
 def runTcCommand(
     command: str,
 ):
-    result = subprocess.run(command.split())
-    if result.returncode > 0:
-        print("Issue running TC!")
-        print(result.args)
-        print(result.stderr)
-        print("--------------------------")
+    if command:
+        result = subprocess.run(command.split())
+        if result.returncode > 0:
+            print("Issue running TC!")
+            print(result.args)
+            print(result.stderr)
+            print("--------------------------")
 
 
 if __name__ == "__main__":
